+++ src/commands/createDatapackTemplate/nodes/index.ts
@@ -1,4 +1,3 @@
-import { locale } from '../../../locales';
 import { AddTemplateGenNode } from './AddTemplateGenNode';
 import { CreateTemplateGenNode } from './CreateTemplateGenNode';
 
@@ -10,8 +9,8 @@ export type GenNodes =
     | typeof CreateTemplateGenNode;
 
 export function getGenTypeMap(): Map<string, GenNodes> {
-    const res = new Map<string, GenNodes>();
-    res.set(locale('create-datapack-template.add'), AddTemplateGenNode);
-    res.set(locale('create-datapack-template.create'), CreateTemplateGenNode);
-    return res;
+    return new Map<string, GenNodes>([
+        ['create-datapack-template.add', AddTemplateGenNode],
+        ['create-datapack-template.create', CreateTemplateGenNode]
+    ]);
 }