+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# [v2.1.0](https://github.com/ChenCMD/MC-Datapack-Utility/compare/v2.0.5...v2.1.0) (2021-12-06)
+
+## ✨ New Features
+- [`3ca8d21`](https://github.com/ChenCMD/MC-Datapack-Utility/commit/3ca8d21)  Add append.addFirst to customTemplate
+
+# [v2.0.5](https://github.com/ChenCMD/MC-Datapack-Utility/compare/v2.0.4...v2.0.5) (2021-09-17)
+
+## 🐛 Bug Fixes
+- [`05feb6f`](https://github.com/ChenCMD/MC-Datapack-Utility/commit/05feb6f)  Revert &quot;🩹 fix locale involved core system&quot; 
+- [`8eaffa9`](https://github.com/ChenCMD/MC-Datapack-Utility/commit/8eaffa9)  fix build
+
+# [v2.0.4](https://github.com/ChenCMD/MC-Datapack-Utility/compare/v2.0.3...v2.0.4) (2021-09-01)
+
 # [v2.0.3](https://github.com/ChenCMD/MC-Datapack-Utility/compare/v2.0.2...v2.0.3) (2021-05-07)
 
 ## 🔒 Security Issues